requirements.txt
h5py==2.9.0
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
matplotlib==3.1.0
MedPy==0.4.0
numpy==1.16.4
opencv-python==4.1.0.25
pandas==0.24.2
Pillow==6.0.0
ptyprocess==0.6.0
pydicom==1.2.2
pymediainfo==4.0
PyQt5==5.12.2
PyQt5-sip==4.19.17
pyradiomics==2.2.0
pytz==2019.1
scikit-image==0.15.0
scikit-learn==0.21.2
scipy==1.3.0
SimpleITK==1.2.0
six==1.12.0
tensorboard==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
truecolor==1.0b2
tensorflow-gpu==1.12.0
tensorflow==1.12.0